% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{address_uk}
\alias{address_uk}
\title{UK addresses.}
\format{A data frame with 5 variables: \code{postcode}, \code{country},
  \code{primary_care_trust}, \code{longitude} and \code{latitude}.}
\usage{
address_uk
}
\description{
A dataset with 10,000 UK addresses extracted from code{\link{extract_address}}, which
uses an API \code{\link[PostcodesioR:random_postcode]{random_postcode}} to sample a
real UK address from \url{https://api.postcodes.io/random/postcodes}.
}
\keyword{datasets}
