% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adult}
\alias{adult}
\title{Adult dataset.}
\format{A data frame with 13 variables: \code{age}, \code{workclass},
  \code{marital_status}, \code{occupation}, \code{relationship},
  \code{race}, \code{sex}, \code{capital_gain}, \code{capital_loss},
  \code{hours_per_week}, \code{native_country} and \code{income}.}
\usage{
adult
}
\description{
The Adult dataset was extracted from the US Census database in 1994; it contains
48,842 individual records with 13 personal variables. It is often used as a
prediction task to determine whether a person makes over $50,000 a year given
personal information.
}
\keyword{datasets}
