% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acquire_error.R
\name{check_swap_char}
\alias{check_swap_char}
\title{Check if two strings are the same after we swaped the position of two letters.}
\usage{
check_swap_char(string1, string2)
}
\arguments{
\item{string1}{A string.}

\item{string2}{A string.}
}
\value{
It returns TRUE if two strings are the same after we swaped the
    position of two letters, otherwise, it returns FALSE.
}
\description{
\code{check_swap_char} check if two strings are the same after we swaped the
     position of two letters.
}
