% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acquire_error.R
\name{diff_two_strings}
\alias{diff_two_strings}
\title{Find all letters in \code{string1} which are not in \code{string2}.
\code{diff_two_strings} is adopted from package vecsets function vsetdiff,
    it returns all letters in \code{string1} which are not in \code{string2}.}
\usage{
diff_two_strings(string1, string2, multiple = TRUE)
}
\arguments{
\item{string1}{A string.}

\item{string2}{A string.}

\item{multiple}{A logical variable with a default of TRUE. If multiple is TRUE, it
will non-unique letters, otherwise, only unique letters.}
}
\description{
Find all letters in \code{string1} which are not in \code{string2}.
\code{diff_two_strings} is adopted from package vecsets function vsetdiff,
    it returns all letters in \code{string1} which are not in \code{string2}.
}
