% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ocr_error.R
\name{do_ocr_replacement}
\alias{do_ocr_replacement}
\title{Replace a string with its ocr error.}
\usage{
do_ocr_replacement(s, where, orgpat, newpat)
}
\arguments{
\item{s}{A string.}

\item{where}{A string. The location of the pat, it can be one of: 'ALL','START','END','MIDDLE'.}

\item{orgpat}{A string. The original pat.}

\item{newpat}{A string. The new pat.}
}
\value{
It returns a new pat.
}
\description{
\code{do_ocr_replacement} replace a string with its ocr error.
}
