% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_typo_error.R
\name{do_typo_replacement}
\alias{do_typo_replacement}
\title{Replace a string with its typo error.}
\usage{
do_typo_replacement(s)
}
\arguments{
\item{s}{A string.}
}
\value{
It returns a new pat.
}
\description{
\code{do_ocr_replacement} replace a string with its typo error.
}
