% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{firstname_uk_variant}
\alias{firstname_uk_variant}
\title{First name variants in the UK.}
\format{A data frame with 3 variables:
\describe{
\item{\code{forename}}{the reference name}
\item{\code{forename2}}{the variant of the \code{forename}}
\item{\code{freq}}{probability of entering \code{forename2} as a variant of \code{forename}}.
}}
\usage{
firstname_uk_variant
}
\description{
A record of first name variants in the UK, provided by ONS.
}
\keyword{datasets}
