% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.R
\name{gen_dob}
\alias{gen_dob}
\title{Generate a record of date of birth.}
\usage{
gen_dob(start = "1900-01-01", end = "2020-01-01")
}
\arguments{
\item{start}{A Date variable with a default of '1900-01-01'.}

\item{end}{A Date variable with a default of '2020-01-01'.}
}
\value{
The output is a record of date of birth in Date format between
  1900-01-01 and 2020-01-01. If \code{start} is given, the return date
  will be between the updated start date and 2020-01-01. If \code{end}
  is also given, the return date will be between the updated start date
  and updated end date.
}
\description{
\code{gen_dob} randomly return a record of date of birth.
}
\examples{
gen_dob()
gen_dob(start = "1995-01-01")
gen_dob(end = "2000-01-01")
gen_dob(start = "1909-01-01", end = "2000-01-01")

}
