% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhsid.R
\name{gen_nhsid}
\alias{gen_nhsid}
\title{Generate a random nhsid.}
\usage{
gen_nhsid()
}
\value{
The output is string with 10 numbers.
}
\description{
\code{gen_nhsid} randomly return a 10-digit nhsid that is generated
    following the Modulus 11 Algorithm;
}
\examples{
gen_nhsid()

}
