% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_generic_error.R
\name{get_transformation_del}
\alias{get_transformation_del}
\title{Delete a character randomly.}
\usage{
get_transformation_del(string)
}
\arguments{
\item{string}{A string.}
}
\value{
It returns the \code{string} with one of the characters was randomly
   deleted. It also comes with the change log of the transformation.
}
\description{
\code{get_transformation_del} randomly delete a character of a string.
}
\examples{
get_transformation_del('how are you?')

}
