% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_generic_error.R
\name{get_transformation_insert}
\alias{get_transformation_insert}
\title{Insert a character/digit/space/symbol randomly.}
\usage{
get_transformation_insert(string)
}
\arguments{
\item{string}{A string.}
}
\value{
It returns the \code{string} with an additional character/digit/space/symbol.
    It also comes with the change log of the transformation.
}
\description{
\code{get_transformation_del} randomly insert a character/digit/space/symbol
    a string.
}
\examples{
get_transformation_insert('how are you?')

}
