% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_generic_error.R
\name{get_transformation_name_variant}
\alias{get_transformation_name_variant}
\title{Randomly assign a name to its variant.}
\usage{
get_transformation_name_variant(string)
}
\arguments{
\item{string}{A name string.}
}
\value{
It returns the name variant of \code{string} together with the change log
    of the transformation. If no name variant was recorded in the database, it
    returns the same name \code{string} with a note of 'no recorded variants'.
}
\description{
\code{get_transformation_name_variant} randomly assign a name to its variant. The
    name variant databases are extracted from Febrl.
}
\examples{
get_transformation_name_variant("ed")
get_transformation_name_variant("shelly")
get_transformation_name_variant("MORRIS")

}
