% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pho_error.R
\name{get_transformation_pho}
\alias{get_transformation_pho}
\title{Encode phonetic error to a string.}
\usage{
get_transformation_pho(string)
}
\arguments{
\item{string}{A string.}
}
\value{
It returns the \code{string} with a randomly assgined phonetic error following
   rules extracted in the pho_rules dataset. It also comes with the change log of the
   transformation.
}
\description{
\code{get_transformation_pho} randomly assign a Phonetic error to a string.
    This function was converted from the Python code in Febrl (developed by
    Agus Pudjijono in 2008, refers to reference \url{https://link.springer.com/chapter/10.1007/978-3-642-01307-2_47}.
}
\examples{
get_transformation_pho('how are you?')

}
