% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_generic_error.R
\name{get_transformation_trans_date}
\alias{get_transformation_trans_date}
\title{Transpose the position of day and month.}
\usage{
get_transformation_trans_date(date)
}
\arguments{
\item{date}{A Date variable.}
}
\value{
The output is the transposition of day and month of \code{date}
   and the change log of the transposition. If the day of \code{date}
   is greater than 12, the transposition will fail and return the same
   \code{date} with a log saying "cannot transposte due to day >12".
}
\description{
\code{get_transformation_trans_date} transpose the position of day and month of a
    Date format variable.
}
\examples{
get_transformation_trans_date("1995-01-11")
get_transformation_trans_date("1995-01-13")

}
