% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lastname_uk}
\alias{lastname_uk}
\title{Last names in UK,}
\format{A data frame with 2 variables: \code{surname} and \code{freq}.}
\usage{
lastname_uk
}
\description{
UK last name dataset was provided by the ONS. The frenquency was calculated
by the number of uses in each name divided by the number of the population
within the dataset.
}
\keyword{datasets}
