% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lastname_uk_variant}
\alias{lastname_uk_variant}
\title{Last name variants in the UK.}
\format{A data frame with 3 variables:
\describe{
\item{\code{lastname1}}{the reference name}
\item{\code{lastname2}}{the variant of the \code{lastname1}}
\item{\code{freq}}{probability of entering \code{lastname2} as a variant of \code{lastname1}}.
}}
\usage{
lastname_uk_variant
}
\description{
A record of last name variants in the UK, provided by ONS.
}
\keyword{datasets}
