% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ocr_rules}
\alias{ocr_rules}
\title{Look up table of Optical Character Recognition (OCR) errors.}
\format{A data frame with 3 variables:
\describe{
\item{\code{postion}}{the position of the error within a string}
\item{\code{orgpat}}{the original pat}
\item{\code{newpat}}{the error pat of \code{orgpat} that is
    misrecogonised by the OCR system}.
}}
\usage{
ocr_rules
}
\description{
A list of OCR errors that may happen, provided by Febrl.
}
\keyword{datasets}
