% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pho_rules}
\alias{pho_rules}
\title{Look up table of phonetic errors.}
\format{A data frame with 7 variables:
\describe{
\item{\code{where}}{the position of the error within a string, can be one of: 'ALL','START','END','MIDDLE'}
\item{\code{orgpat}}{the original pat}
\item{\code{newpat}}{the error pat of \code{orgpat} that is misheard}
\item{\code{precond}}{pre-condition (default 'None') can be 'V' for vowel or 'C' for consonant}
\item{\code{postcond}}{post-condition (default 'None') can be 'V' for vowel or 'C' for consonant}
\item{\code{existcond}}{exist-condition (default 'None')}
\item{\code{startcond}}{start-condition (default 'None')}.
}}
\usage{
pho_rules
}
\description{
A list of phonetic errors that may happen, provided by Febrl.
}
\keyword{datasets}
