% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_bn.R
\name{plot_bn}
\alias{plot_bn}
\title{Plot the BN structure.}
\usage{
plot_bn(structure, ht = "400px")
}
\arguments{
\item{structure}{A string of the relationships between variables from
\code{\link[bnlearn:model string utilities]{modelstring}}.}

\item{ht}{The height of the plot.}
}
\value{
The output is a plot of the Bayesian Network structure.
}
\description{
\code{plot_bn} generates a plot of the Bayesian Network structure.
}
\examples{
adult_data <- split_data(adult[1:100,], 70)
bn_learn = gen_bn_learn(adult_data$training_set, 'hc')
plot_bn(bn_learn$structure)

}
