% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pho_error.R
\name{slavo_germanic}
\alias{slavo_germanic}
\title{Detect if it has slavo transformation.}
\usage{
slavo_germanic(str)
}
\arguments{
\item{str}{A string.}
}
\value{
It returns 1 or 0.
}
\description{
\code{slavo_germanic} detect if a string has slavo transformation.
}
