% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data.R
\name{split_data}
\alias{split_data}
\title{Split the data into a training_set and a testing_set.}
\usage{
split_data(dataset, training_percentage)
}
\arguments{
\item{dataset}{A data frame of the dataset.}

\item{training_percentage}{A number between 0 and 100 indicating the
percertage of the training dataset.}
}
\value{
A list with two data frame: training_set and testing_set.
}
\description{
\code{split_data} split the data into a training_set
    and a testing_set based on the \code{training_percentage}.
}
\examples{
adult_data <- split_data(adult[1:100,], 70)

}
